+++ src/lib/jp2/tile/TileComponent.h
@@ -54,8 +54,8 @@ struct TileComponent : public grk_rect_u32 {
 	 ISparseBuffer* getSparseBuffer();
 
 	Resolution *resolutions; /* resolutions information */
-	uint32_t numresolutions; /* number of resolution levels */
-	uint32_t resolutions_to_decompress; /* number of resolutions level to decompress (at max)*/
+	uint8_t numresolutions; /* number of resolution levels */
+	uint8_t resolutions_to_decompress; /* number of resolutions level to decompress (at max)*/
 #ifdef DEBUG_LOSSLESS_T2
 	Resolution* round_trip_resolutions;  /* round trip resolution information */
 #endif
